chore: Move App.tsx from /src to src/app/
- Moved the App.tsx and associated files to /src/app to align with a modern React feature-based folder structure.
