+++ frontend/src/app/App.tsx
@@ -1,13 +1,13 @@
 import './App.css';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import HomePage from './features/home/Home';
-import RouteLayout from './pages/RootLayout';
-import ErrorPage from './pages/ErrorPage';
-import BookDetails from './pages/BookDetails';
-import { loader as bookLoader } from './pages/BookDetails';
-import BookSearch from './features/shelves/BookSearch';
-import BookShelfContextProvider from './store/BookShelfContext';
-import MyBookList from './features/booklist/MyBooklist';
+import HomePage from '../features/home/Home';
+import RouteLayout from '../pages/RootLayout';
+import ErrorPage from '../pages/ErrorPage';
+import BookDetails from '../pages/BookDetails';
+import { loader as bookLoader } from '../pages/BookDetails';
+import BookSearch from '../features/shelves/BookSearch';
+import BookShelfContextProvider from '../store/BookShelfContext';
+import MyBookList from '../features/booklist/MyBooklist';
 
 
 const router = createBrowserRouter([
